Increasing EyeDistance with Input and other QOL changes
The value for the EyeDistance in each star can be changed with the use of the up key. On top of this, the Screen is a reference now, meaning it cannot be null! An error check was implemented in main.cpp so that the programme doesn't run unless Screen is initialised