More efficient blitting
the AlphaBlit() function was rewritten to be more efficient, not doing any floating point arithmetic, only drawing if there is an alpha. It does no calculation if the alpha is 255 (completely opaque) and only calculates the blended alpha channel if there is an alpha channel. Bitshifting was used as a more efficient division method.