Proper Graphics Class
A graphics class was created to deal with the graphics in the game engine, most importantly the m_textureMap, which is an unordered map that holds pointers to sprites and a string identifier.

Changes were made across the board to adopt this new way of identifying textures. This cleans up the interfaces of the classes, and ensures that Texture is 'Black Box'.