Basic Gameplay
Bricks are now entities, allowing for collision detection. They can be hit and destroyed, with green bricks taking 2 hits from the ball to destroy them.

TODO: Sometimes collision can be a bit iffy, I need to offset the paddle's collision and see what is happening when the ball gets trapped at the top.