Animation System
An animation system was created with an Animator class, which controls the current animation, and an Animation class which holds information about the animation, such as the sprite identifiers and the state of the animation (whether it is playing, paused or completed). The Entity class was adapted so that each Entity has an Animator.

The Ball and Brick classes were removed, since they are for a previous game