Enemies
There are now enemies on screen that have a walking animation. They turn around whenever they hit a block to the left or right of them. They also die when the player collides with them. Enemies have a boolean to say whether they should fall off edges or whether they should turn around and stay on their current platform. Enemies can collide with the player and they play a death animation when they do so.

Animations now have a boolean to say if they are looping (such as a walk animation) or not (like a death animation) as well as a parameter for duration, which is defaulted to 0.1s (100ms)

There is now a direction enum for entities, which is much more robust than the vector I was using before.