Player Damage, Lives, and Respawning
If the player hits an enemy and he isn't powered up, he dies, and gets put back to the beginning of the level. When he dies, he loses a life and that is shown on the new UI at the top of the screen. If the player falls off the bottom of the screen, he dies and a life is taken away - the same applies for if he touches spikes. If the player has a powerup and he touches an enemy, he loses the power up and gains 2 seconds of immunity, shown by the sprite changing opacity. A sound effect was added that plays upon player death