Boss Battle
There is now a yellow alien at the end of the last level that acts as a basic boss fight. It waits until the player is within range and then, limited by the bridge at the end of the level, moves towards him, jumping and shooting fireballs at the player to try damage them.